+++ vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
       entry: {
         index: "src/index.ts",
         schemas: "src/schemas.ts",
+        deprecations: "src/deprecations.ts",
         "vite/index": "src/vite/index.ts",
       },
       formats: ["es"],